HW_7/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
httpx>=0.27.0
